requirements.txt
kafka-python
apache-flink
prefect
psycopg2-binary
azure-storage-blob
python-dotenv